SAM working build
