+++ backend/.aws-sam/build.toml
@@ -1,0 +1,11 @@
+# This file is auto generated by SAM CLI build command
+
+[function_build_definitions]
+[function_build_definitions.0b3c3375-7609-4ad6-83a6-d70a7e808856]
+codeuri = "/Users/abe/git/permaximize/backend"
+runtime = "nodejs14.x"
+source_md5 = ""
+packagetype = "Zip"
+functions = ["getAllItemsFunction", "getByIdFunction", "putItemFunction"]
+
+[layer_build_definitions]
